refactor: separate domain package
