+++ src/main/kotlin/racingcar/domain/car/CarAccelerator.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 interface CarAccelerator {
