+++ src/main/kotlin/racingcar/domain/car/CarAction.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 enum class CarAction(val distance: Int) {
