+++ src/main/kotlin/racingcar/domain/car/CarNames.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 data class CarNames(private val names: List<String>) {
