+++ src/main/kotlin/racingcar/domain/racing/RacingCarAccelerator.kt
@@ -1,4 +1,7 @@
-package racingcar.domain
+package racingcar.domain.racing
+
+import racingcar.domain.car.CarAccelerator
+import racingcar.domain.car.CarAction
 
 class RacingCarAccelerator : CarAccelerator {
