+++ src/test/kotlin/racingcar/domain/car/CarPositionTest.kt
@@ -1,10 +1,12 @@
-package racingcar.domain
+package racingcar.domain.car
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import racingcar.domain.car.CarAction
+import racingcar.domain.car.CarPosition
 import java.util.stream.Stream
 
 internal class CarPositionTest {
