+++ src/test/kotlin/racingcar/domain/car/TestCarAccelerator.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 object MoveCarAccelerator : CarAccelerator {
