+++ src/test/kotlin/racingcar/domain/racing/RacingCarGameTest.kt
@@ -1,9 +1,12 @@
-package racingcar.domain
+package racingcar.domain.racing
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import racingcar.domain.car.MoveCarAccelerator
+import racingcar.domain.car.Car
+import racingcar.domain.car.Cars
 import java.util.stream.Stream
 
 internal class RacingCarGameTest {
